requirements.txt
altgraph==0.17.4
contourpy==1.2.0
cycler==0.12.1
EasyProcess==1.1
entrypoint2==1.1
fonttools==4.44.0
importlib-metadata==6.8.0
importlib-resources==6.1.0
keyboard==0.13.5
kiwisolver==1.4.5
matplotlib==3.8.1
MouseInfo==0.1.3
mss==9.0.1
numpy==1.26.1
packaging==23.2
pefile==2023.2.7
Pillow==10.1.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pyinstaller==6.1.0
pyinstaller-hooks-contrib==2023.10
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
pyserial==3.5
python-dateutil==2.8.2
pytweening==1.0.7
pywin32-ctypes==0.2.2
six==1.16.0
zipp==3.17.0
